backend/verification/services/facial_service_simple.py
import logging
import time
from typing import Tuple, Dict, Any
import base64
from io import BytesIO
from PIL import Image
import os

logger = logging.getLogger(__name__)

class SimpleFacialService:
    """
    Simple facial verification service that uses basic image analysis
    and can be enhanced later with proper ML models
    """
    
    def __init__(self):
        self.min_face_size = 100  # Minimum face size in pixels
        self.liveness_threshold = 0.7
        logger.info("Simple facial service initialized")

    def verify_liveness(self, image_path: str) -> Tuple[bool, float, Dict[str, Any]]:
        """
        Basic liveness verification using image metadata and basic checks
        In production, replace with proper liveness detection
        """
        try:
            # Basic image validation
            image_info = self._analyze_image_basic(image_path)
            
            # Simulate liveness check (replace with actual implementation)
            liveness_score = self._simulate_liveness_check(image_info)
            is_live = liveness_score >= self.liveness_threshold
            
            details = {
                'image_quality': image_info.get('quality_score', 0.0),
                'file_size': image_info.get('file_size', 0),
                'dimensions': image_info.get('dimensions', (0, 0)),
                'checks_performed': ['file_validation', 'basic_quality']
            }
            
            logger.info(f"Basic liveness check completed. Score: {liveness_score}")
            
            return is_live, liveness_score, details
            
        except Exception as e:
            logger.error(f"Liveness verification failed: {str(e)}")
            return False, 0.0, {'error': str(e)}

    def compare_faces_basic(self, id_image_path: str, selfie_image_path: str) -> Tuple[bool, float, Dict[str, Any]]:
        """
        Basic face comparison - placeholder for actual facial recognition
        In production, integrate with proper face matching service
        """
        try:
            # For now, simulate face matching
            # In production, integrate with:
            # - AWS Rekognition
            # - Google Vision API
            # - Azure Face API
            # - Or proper on-prem ML model
            
            id_image_info = self._analyze_image_basic(id_image_path)
            selfie_image_info = self._analyze_image_basic(selfie_image_path)
            
            # Simulate matching logic (replace with actual face matching)
            match_score = self._simulate_face_matching(id_image_info, selfie_image_info)
            is_match = match_score >= 0.7  # Adjust threshold as needed
            
            details = {
                'id_image_quality': id_image_info.get('quality_score', 0.0),
                'selfie_image_quality': selfie_image_info.get('quality_score', 0.0),
                'method': 'simulated_matching',
                'note': 'Replace with actual face matching service'
            }
            
            logger.info(f"Basic face comparison completed. Score: {match_score}")
            
            return is_match, match_score, details
            
        except Exception as e:
            logger.error(f"Face comparison failed: {str(e)}")
            return False, 0.0, {'error': str(e)}

    def _analyze_image_basic(self, image_path: str) -> Dict[str, Any]:
        """Basic image analysis without heavy dependencies"""
        try:
            with Image.open(image_path) as img:
                width, height = img.size
                file_size = os.path.getsize(image_path)
                
                # Basic quality score based on dimensions and file size
                quality_score = min(
                    (width * height) / (1000 * 1000),  # Normalize by 1MP
                    file_size / (500 * 1024)  # Normalize by 500KB
                )
                
                return {
                    'dimensions': (width, height),
                    'file_size': file_size,
                    'quality_score': min(quality_score, 1.0),
                    'format': img.format
                }
                
        except Exception as e:
            logger.error(f"Image analysis failed: {str(e)}")
            return {'quality_score': 0.0, 'error': str(e)}

    def _simulate_liveness_check(self, image_info: Dict[str, Any]) -> float:
        """Simulate liveness detection (replace with actual implementation)"""
        # Basic checks that don't require heavy ML
        quality_score = image_info.get('quality_score', 0.0)
        file_size = image_info.get('file_size', 0)
        width, height = image_info.get('dimensions', (0, 0))
        
        score = 0.0
        
        # Check image dimensions (reasonable for face photos)
        if width >= 300 and height >= 300:
            score += 0.3
        
        # Check file size (not too small, not too large)
        if 50 * 1024 <= file_size <= 5 * 1024 * 1024:  # 50KB to 5MB
            score += 0.3
        
        # Check image quality
        if quality_score >= 0.5:
            score += 0.4
        
        return min(score, 1.0)

    def _simulate_face_matching(self, id_info: Dict[str, Any], selfie_info: Dict[str, Any]) -> float:
        """Simulate face matching (replace with actual face recognition)"""
        # For now, return a reasonable score if both images meet quality thresholds
        id_quality = id_info.get('quality_score', 0.0)
        selfie_quality = selfie_info.get('quality_score', 0.0)
        
        if id_quality >= 0.6 and selfie_quality >= 0.6:
            return 0.85  # Simulated good match
        else:
            return 0.3  # Simulated poor match

    def validate_facial_image(self, image_path: str) -> Tuple[bool, str]:
        """Basic facial image validation"""
        try:
            with Image.open(image_path) as img:
                width, height = img.size
                
                # Check minimum dimensions
                if width < 300 or height < 300:
                    return False, "Image too small. Minimum 300x300 pixels required."
                
                # Check aspect ratio (roughly expected for face photos)
                aspect_ratio = width / height
                if aspect_ratio < 0.6 or aspect_ratio > 1.4:
                    return False, "Image aspect ratio not suitable for face photo."
                
                # Check file size
                file_size = os.path.getsize(image_path)
                if file_size > 5 * 1024 * 1024:  # 5MB
                    return False, "Image file too large. Maximum 5MB allowed."
                
                return True, "Image validation passed"
                
        except Exception as e:
            return False, f"Image validation failed: {str(e)}"